+++ pkg/resultstore/resultstore.go
@@ -1,11 +1,7 @@
 package resultstore
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
-	"io/fs"
 	"time"
 
 	"github.com/iver-wharf/wharf-cmd/pkg/worker"
@@ -64,51 +60,4 @@ type store struct {
 	lastArtifactID uint64
 
 	statusMutex keyedMutex
-}
-
-func (s *store) AddArtifact(stepID uint64, artifactName string) (io.WriteCloser, error) {
-	//artifactID := atomic.AddUint64(&s.lastArtifactID, 1)
-	return nil, nil
-}
-
-func (s *store) resolveLogPath(stepID uint64) string {
-	return fmt.Sprintf("steps/%d/logs.log", stepID)
-}
-
-func (s *store) resolveStatusPath(stepID uint64) string {
-	return fmt.Sprintf("steps/%d/status.json", stepID)
-}
-
-func (s *store) resolveArtifactListMetaPath(stepID uint64) string {
-	return fmt.Sprintf("steps/%d/artifacts.json", stepID)
-}
-
-func (s *store) resolveArtifactPath(stepID uint64, artifactID uint64) (string, error) {
-	listMeta, err := s.readArtifactListMeta(stepID)
-	if err != nil {
-		return "", err
-	}
-	for _, meta := range listMeta.Artifacts {
-		if meta.ArtifactID == artifactID {
-			return meta.Path, nil
-		}
-	}
-	return "", fmt.Errorf("step %d: artifact by ID %d not found in artifacts.json", stepID, artifactID)
-}
-
-func (s *store) readArtifactListMeta(stepID uint64) (ArtifactListMeta, error) {
-	file, err := s.fs.OpenRead(s.resolveArtifactListMetaPath(stepID))
-	if errors.Is(err, fs.ErrNotExist) {
-		return ArtifactListMeta{}, nil
-	}
-	if err != nil {
-		return ArtifactListMeta{}, fmt.Errorf("step %d: read artifact.json: %w", stepID, err)
-	}
-	defer file.Close()
-	decoder := json.NewDecoder(file)
-	var listMeta ArtifactListMeta
-	if err := decoder.Decode(&listMeta); err != nil {
-		return ArtifactListMeta{}, err
-	}
-	return listMeta, nil
 }
